- Move setup.py to project root directory     - it looks like the GitHub workflow expects this location - Modify test imports to account for new paths
